pyproject.toml
[project]
name = "klublepszezycie"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "flask>=3.1.2",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "flower>=2.0.1",
    "gunicorn>=23.0.0",
    "numpy>=2.3.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "redis>=6.4.0",
    "unidecode>=1.4.0",
    "user-agents==2.2.0",
    "werkzeug>=3.1.3",
    "xlrd<2.0",
    "xlutils>=2.0.0",
    "xlwt>=1.3.0",
]
